Add LZ4 compression and increase data sizes to put extra pressure on cluster
In practice, this is silly because random byte data doesn't compress
well at all. The compressed data is almost always larger than the source
data, making the exercise pointless and waste of CPU and memory
resources.

However, it serves as a good example of using the compressor which would
work much better if the data was text or heavy JSON payloads.